Revert 'unneeded' changes
