+++ WebAPI.OData.EFCore/Attributes/AllowApplyAttribute.cs
@@ -8,6 +8,11 @@ namespace WebAPI.OData.EFCore.Attributes
 {
     public class AllowApplyAttribute : EnableQueryAttribute
     {
+        public AllowApplyAttribute()
+        {
+            MaxExpansionDepth = 10;
+        }
+
         public override IQueryable ApplyQuery(IQueryable queryable, ODataQueryOptions queryOptions)
         {
             var ignoreQueryOptions =
